+++ src/Service/Nuvei.php
@@ -122,7 +122,6 @@ class Nuvei
     private $restApiProdUrl          = 'https://secure.safecharge.com/ppp/api/v1/';
     private $saveLogs                = true;
     private $sandboxMode             = true;
-    private $nuveiSourceApplication  = 'SHOPWARE_PLUGIN';
     
     private $devices     = array('iphone', 'ipad', 'android', 'silk', 'blackberry', 'touch', 'linux', 'windows', 'mac');
     private $browsers    = array('ucbrowser', 'firefox', 'chrome', 'opera', 'msie', 'edge', 'safari', 'blackberry', 'trident');
@@ -363,7 +362,7 @@ class Nuvei
                 'timeStamp'         => $time,
                 'deviceDetails'     => $this->getDeviceDetails(),
                 'webMasterId'       => $webMasterId,
-                'sourceApplication' => $this->nuveiSourceApplication,
+                'sourceApplication' => $this->getSourceApplication(),
                 'url'               => $site_url . '/nuvei_dmn/', // a custom parameter for the checksum
                 'urlDetails'        => [
                     'notificationUrl'   => $site_url . '/nuvei_dmn/',
@@ -514,6 +513,11 @@ class Nuvei
         }
         
         return '';
+    }
+    
+    public function getSourceApplication()
+    {
+        return 'SHOPWARE_PLUGIN';
     }
     
     /**